Move listen Uri to configuration
As requested listen Uri is now a part of configuration.
